feat: seperate out views and show earned badges before mint
